requirements.txt
torch
numpy
networkx
scipy
scikit-learn
hydra-core
transformers~=4.18.0
omegaconf
tqdm
dgl
pytorch_lightning
apex
gensim
nltk
seaborn
numba
